Update on "add model code to memory profiler"
See summary in https://github.com/pytorch/pytorch/pull/166450


cc ezyang EikanWang jgong5 wenzhe-nrv

[ghstack-poisoned]
